Move InitializationDelegate into iOS Crashes Binding namespace to reduce its visibility (and because it does not belong in Crashes class to begin with)
